+++ src/tool/cpp/basewinrt/base.h
@@ -9,12 +9,12 @@
 #include "base_handle.h"
 #include "base_lock.h"
 #include "base_abi.h"
+#include "base_com_ptr.h"
 #include "base_windows.h"
 #include "base_string.h"
 #include "base_string_input.h"
 #include "base_string_operators.h"
 #include "base_array.h"
-#include "base_com_ptr.h"
 #include "base_weak_ref.h"
 #include "base_agile_ref.h"
 #include "base_error.h"
